Route voting chain submission through the SDK
Make the version-18 SDK lifecycle authoritative for voting chain submission, including Tor dispatch classification, recovery scheduling, cancellation, and confirmation mapping.

+++ test/features/voting/voting_recovery_service_test.dart
@@ -458,7 +458,7 @@ rust_frb_types.DelegationRecoveryView delegationWorkflow({
   required int bundleIndex,
   required String phase,
   String? txHash,
-  int? vanLeafPosition,
+  BigInt? vanLeafPosition,
 }) {
   return rust_frb_types.DelegationRecoveryView(
     bundleIndex: bundleIndex,
